Attempt to refactor GenerateTestsAction classes
