+++ src/main/kotlin/nl/tudelft/ewi/se/ciselab/testgenie/actions/GenerateTestsActionLine.kt
@@ -5,9 +5,6 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.editor.Caret
-import com.intellij.openapi.project.Project
-import com.intellij.openapi.roots.ProjectRootManager
-import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiFile
 import nl.tudelft.ewi.se.ciselab.testgenie.evosuite.Runner
 
@@ -25,29 +22,33 @@ class GenerateTestsActionLine : AnAction() {
      * @param e an action event that contains useful information
      */
     override fun actionPerformed(e: AnActionEvent) {
-        val project: Project = e.project ?: return
+//        val project: Project = e.project ?: return
+
+        val evoSuiteRunner: Runner = createEvoSuiteRunner(e) ?: return
 
         val psiFile: PsiFile = e.dataContext.getData(CommonDataKeys.PSI_FILE) ?: return
         val caret: Caret = e.dataContext.getData(CommonDataKeys.CARET)?.caretModel?.primaryCaret ?: return
 
-        val psiClass: PsiClass = getSurroundingClass(psiFile, caret) ?: return
-        val classFQN = psiClass.qualifiedName ?: return
+//        val psiClass: PsiClass = getSurroundingClass(psiFile, caret) ?: return
+//        val classFQN = psiClass.qualifiedName ?: return
 
         val selectedLine: Int = getSurroundingLine(psiFile, caret)?.plus(1) ?: return // lines in the editor and in EvoSuite are one-based
-        val vFile = e.dataContext.getData(CommonDataKeys.VIRTUAL_FILE) ?: return
-        val fileName = vFile.presentableUrl
-        val modificationStamp = vFile.modificationStamp
+//        val vFile = e.dataContext.getData(CommonDataKeys.VIRTUAL_FILE) ?: return
+//        val fileName = vFile.presentableUrl
+//        val modificationStamp = vFile.modificationStamp
 
-        val projectPath: String = ProjectRootManager.getInstance(project).contentRoots.first().path
-        val projectClassPath = "$projectPath/target/classes/"
+//        val projectPath: String = ProjectRootManager.getInstance(project).contentRoots.first().path
+//        val projectClassPath = "$projectPath/target/classes/"
 
-        log.info("Generating tests for project $projectPath with classpath $projectClassPath")
+//        log.info("Generating tests for project $projectPath with classpath $projectClassPath")
 
-        log.info("Selected class is $classFQN")
-        log.info("Selected line is $selectedLine")
+//        log.info("Selected class is $classFQN")
 
-        Runner(project, projectPath, projectClassPath, classFQN, fileName, modificationStamp).forLine(selectedLine)
-            .runEvoSuite()
+        Logger.getInstance("Test Generation").info("Selected line is $selectedLine")
+
+//        Runner(project, projectPath, projectClassPath, classFQN, fileName, modificationStamp).forLine(selectedLine)
+//            .runEvoSuite()
+        evoSuiteRunner.forLine(selectedLine).runEvoSuite()
     }
 
     /**
@@ -62,7 +63,8 @@ class GenerateTestsActionLine : AnAction() {
         val caret: Caret = e.dataContext.getData(CommonDataKeys.CARET)?.caretModel?.primaryCaret ?: return
         val psiFile: PsiFile = e.dataContext.getData(CommonDataKeys.PSI_FILE) ?: return
 
-        val line: Int = getSurroundingLine(psiFile, caret)?.plus(1) ?: return // lines in the editor and in EvoSuite are one-based
+        val line: Int = getSurroundingLine(psiFile, caret)?.plus(1)
+            ?: return // lines in the editor and in EvoSuite are one-based
 
         e.presentation.isEnabledAndVisible = true
         e.presentation.text = "Generate Tests For Line $line"
